requirements.txt
Bottleneck==1.3.5
certifi==2022.6.15
distro==1.5.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.3
numpy
packaging==21.3
pandas==1.4.3
pip==22.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
setuptools==63.4.1
six==1.16.0
tabula-py==2.3.0
wheel==0.37.1
